docs: fixes image rendering on docs/index.rst
Images in README.md were not being rendered on the docs site due to the
use of HTML image tags, which are incompatible with Pandoc. Conversion
to markdown image tags and the use of additional conversion logic in
generate-docs has resolved this issue.

Changelog: fixed - Fixed image rendering in docs/index.rst