feat: sets a default managedNamespace in values.yaml
The lack of a default managedNamespace for the slurm-bridge admission
controller results in a confusing deployment process when following the
installation guide. A default managedNamespace should be set in
values.yaml in order to ease the deployment of slurm-bridge, and the
documentation should be updated to inform users about setting this
value.

Closes #51

Changelog: added - Added default managedNamespace for slurm-bridge
 admission controller